+++ code/core/src/main/java/org/betonquest/betonquest/quest/action/effect/DeleteEffectActionFactory.java
@@ -3,10 +3,10 @@ package org.betonquest.betonquest.quest.action.effect;
 import org.betonquest.betonquest.api.QuestException;
 import org.betonquest.betonquest.api.instruction.Argument;
 import org.betonquest.betonquest.api.instruction.Instruction;
+import org.betonquest.betonquest.api.instruction.argument.parser.PotionEffectTypeParser;
 import org.betonquest.betonquest.api.quest.action.OnlineActionAdapter;
 import org.betonquest.betonquest.api.quest.action.PlayerAction;
 import org.betonquest.betonquest.api.quest.action.PlayerActionFactory;
-import org.betonquest.betonquest.lib.instruction.argument.DefaultArgument;
 import org.bukkit.potion.PotionEffectType;
 
 import java.util.Collections;
@@ -25,20 +25,8 @@ public class DeleteEffectActionFactory implements PlayerActionFactory {
 
     @Override
     public PlayerAction parsePlayer(final Instruction instruction) throws QuestException {
-        final Argument<List<PotionEffectType>> effects;
-        final boolean any = instruction.bool().getFlag("any", true)
-                .getValue(null).orElse(false);
-        if (!any && instruction.size() > 1 && !instruction.nextElement().startsWith("conditions:")) {
-            effects = instruction.chainForArgument(instruction.current()).parse(type -> {
-                final PotionEffectType effect = PotionEffectType.getByName(type);
-                if (effect == null) {
-                    throw new QuestException("Unknown effect type: " + type);
-                }
-                return effect;
-            }).list().get();
-        } else {
-            effects = new DefaultArgument<>(Collections.emptyList());
-        }
+        final Argument<List<PotionEffectType>> effects = instruction.parse(PotionEffectTypeParser.POTION_EFFECT_TYPE)
+                .list().prefilter("any", Collections.emptyList()).get();
         return new OnlineActionAdapter(new DeleteEffectAction(effects));
     }
 }