+++ code/core/src/main/java/org/betonquest/betonquest/quest/condition/effect/EffectConditionFactory.java
@@ -3,6 +3,7 @@ package org.betonquest.betonquest.quest.condition.effect;
 import org.betonquest.betonquest.api.QuestException;
 import org.betonquest.betonquest.api.instruction.Argument;
 import org.betonquest.betonquest.api.instruction.Instruction;
+import org.betonquest.betonquest.api.instruction.argument.parser.PotionEffectTypeParser;
 import org.betonquest.betonquest.api.quest.condition.OnlineConditionAdapter;
 import org.betonquest.betonquest.api.quest.condition.PlayerCondition;
 import org.betonquest.betonquest.api.quest.condition.PlayerConditionFactory;